Added lock package

+++ lock/ILock.go
@@ -1,9 +1,9 @@
 package lock
 
 type ILock interface {
-	TryAcquireLock(correlationId string, key string, ttl int) (bool, error)
+	TryAcquireLock(correlationId string, key string, ttl int64) (bool, error)
 
-	AcquireLock(correlationId string, key string, ttl int, timeout int) error
+	AcquireLock(correlationId string, key string, ttl int64, timeout int64) error
 
 	ReleaseLock(correlationId string, key string) error
 }
